+++ test/unit/gen_libs/chk_crt_file.py
@@ -90,7 +90,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = None
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("755", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, exe=True, no_print=True)
@@ -109,7 +110,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = self.prt_template4 % (self.f_name)
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("444", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, exe=True, no_print=True)
@@ -131,7 +133,8 @@ class UnitTest(unittest.TestCase):
         err_msg_chk2 = self.prt_template4 % (self.f_name)
         err_msg_chk = "\n".join([err_msg_chk, err_msg_chk2])
         err_msg_chk = err_msg_chk.strip("\n")
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("444", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, exe=True, write=True, no_print=True)
@@ -153,7 +156,8 @@ class UnitTest(unittest.TestCase):
         err_msg_chk2 = self.prt_template4 % (self.f_name)
         err_msg_chk = "\n".join([err_msg_chk, err_msg_chk2])
         err_msg_chk = err_msg_chk.strip("\n")
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("222", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, exe=True, read=True, no_print=True)
@@ -237,7 +241,8 @@ class UnitTest(unittest.TestCase):
         err_msg_chk2 = self.prt_template3 % (self.f_name)
         err_msg_chk = "\n".join([err_msg_chk, err_msg_chk2])
         err_msg_chk = err_msg_chk.strip("\n")
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("111", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, write=True, read=True, no_print=True)
@@ -305,7 +310,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = None
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, write=True, no_print=True)
 
@@ -323,7 +329,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = self.prt_template2 % (self.f_name)
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("444", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, write=True, no_print=True)
@@ -342,7 +349,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = None
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, read=True, no_print=True)
 
@@ -360,7 +368,8 @@ class UnitTest(unittest.TestCase):
         """
 
         err_msg_chk = self.prt_template3 % (self.f_name)
-        open(self.f_name, "a", encoding="UTF-8").close() # pylint:disable=R1732
+        open(                                       # pylint:disable=R1732
+            self.f_name, "a", encoding="UTF-8").close()
         os.chmod(self.f_name, int("333", 8))
         status, err_msg = gen_libs.chk_crt_file(
             self.f_name, read=True, no_print=True)